+++ spec/types/model_spec.rb
@@ -208,7 +208,7 @@ describe Attributor::Model do
         it 'and sets them in loaded format onto the instance attributes' do
           Chicken.should_receive(:load).with(attributes_data).and_call_original
           attributes_data.keys.each do |attr_name|
-            Chicken.attributes[attr_name].should_receive(:load).with(attributes_data[attr_name]).and_call_original
+            Chicken.attributes[attr_name].should_receive(:load).with(attributes_data[attr_name],instance_of(Array)).and_call_original
           end
           subject.age.should be(1)
           subject.email.should be(attributes_data[:email])
@@ -220,7 +220,7 @@ describe Attributor::Model do
           it 'and sets them in loaded format onto the instance attributes' do
             Chicken.should_receive(:load).with(attributes_data).and_call_original
             attributes_hash.keys.each do |attr_name|
-              Chicken.attributes[attr_name].should_receive(:load).with(attributes_hash[attr_name]).and_call_original
+              Chicken.attributes[attr_name].should_receive(:load).with(attributes_hash[attr_name],instance_of(Array)).and_call_original
             end
             subject.age.should be(1)
             subject.email.should == attributes_hash[:email]